# docs/api-spec/v0.2.0/openapi.yaml
openapi: 3.1.0
info:
  title: LakeMind API
  version: 1.0.0
  description: |
    LakeMind v0.2.0 REST API — 统一认知资产管理与受控 Job Runtime。
    所有端点（除 /health）需要 Bearer Token 认证。
    X-Tenant-Id / X-Agent-Id / X-Scopes 不再由调用方声明，由 Token 解析得到。
  contact:
    name: LakeMind Team
servers:
  - url: http://localhost:10823
    description: Local development
  - url: http://lakemind-server:10823
    description: Docker internal
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/assets:
    post:
      tags: [Assets]
      summary: 创建资产
      operationId: createAsset
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '422':
          $ref: '#/components/responses/ValidationError'
    get:
      tags: [Assets]
      summary: 资产列表
      operationId: listAssets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/AssetType'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AssetStatus'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAsset'
...
  /api/v1/knowledge:
    post:
      tags: [Knowledge]
      summary: 摄入知识（异步 Operation）
      operationId: ingestKnowledge
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestKnowledgeRequest'
      responses:
        '202':
          description: 摄入已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncOperationResponse'
    get:
      tags: [Knowledge]
      summary: 知识库列表
      operationId: listKnowledge
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseKnowledgeBase'
...
  /api/v1/skills:
    post:
      tags: [Skills]
      summary: 注册 Skill（DRAFT）
      operationId: registerSkill
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterSkillRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
    get:
      tags: [Skills]
      summary: Skill 列表
      operationId: listSkills
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseSkill'
...
  /api/v1/skills/{name}/{version}:
    get:
      tags: [Skills]
      summary: Skill 详情
      operationId: getSkill
      parameters:
        - $ref: '#/components/parameters/SkillName'
        - $ref: '#/components/parameters/SkillVersion'
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/skills/{name}/{version}/publish:
    post:
      tags: [Skills]
      summary: 发布 Skill（不可变）
      operationId: publishSkill
      parameters:
        - $ref: '#/components/parameters/SkillName'
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/skills/{name}/{version}/revoke:
    post:
      tags: [Skills]
      summary: 撤销 Skill
      operationId: revokeSkill
      parameters:
        - $ref: '#/components/parameters/SkillName'
        - $ref: '#/components/parameters/SkillVersion'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: 撤销成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'

  # ────────────── Memory ──────────────
  /api/v1/memories:
    post:
      tags: [Memory]
      summary: 添加 Memory
      operationId: addMemory
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemoryRequest'
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
    get:
      tags: [Memory]
      summary: Memory 列表
      operationId: listMemories
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseMemory'
...
  /api/v1/jobs:
    post:
      tags: [Jobs]
      summary: 提交 Job
      operationId: submitJob
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitJobRequest'
      responses:
        '201':
          description: 提交成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobRun'
    get:
      tags: [Jobs]
      summary: Job 列表
      operationId: listJobs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseJobRun'
...
  /api/v1/operations:
    post:
      tags: [Operations]
      summary: 创建 Operation
      operationId: createOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperationRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
    get:
      tags: [Operations]
      summary: Operation 列表
      operationId: listOperations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OperationStatus'
        - name: type
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseOperation'
...
  /api/v1/models:
    post:
      tags: [Models]
      summary: 创建模型定义
      operationId: createModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModelRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDefinition'
    get:
      tags: [Models]
      summary: 模型列表
      operationId: listModels
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDefinition'
...
  /api/v1/security/secrets:
    post:
      tags: [Security]
      summary: 创建 Secret
      operationId: createSecret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecretRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretRef'
    get:
      tags: [Security]
      summary: Secret 元数据列表
      operationId: listSecrets
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecretMetadata'
...
  /api/v1/audit:
    get:
      tags: [Audit]
      summary: 查询审计事件
      operationId: queryAudit
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: event_type
          in: query
          schema:
            type: string
        - name: principal
          in: query
          schema:
            type: string
        - name: resource
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 审计事件列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseAuditEvent'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: 幂等键（24h 内相同 Principal + 相同 Key 返回缓存响应）
...
    SkillName:
      name: name
      in: path
      required: true
      schema:
        type: string
    SkillVersion:
      name: version
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: 校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - AUTHENTICATION_FAILED
                - TOKEN_REVOKED
                - PERMISSION_DENIED
                - TENANT_SCOPE_VIOLATION
                - RESOURCE_NOT_FOUND
                - ASSET_NOT_READY
                - ASSET_DEGRADED
                - ASSET_FAILED
                - SKILL_NOT_PUBLISHED
                - SKILL_VERSION_IMMUTABLE
                - JOB_RESOURCE_DENIED
                - JOB_NOT_FOUND
                - MODEL_DEPLOYMENT_UNAVAILABLE
                - EMBEDDING_SPACE_MISMATCH
                - CONFIG_REVISION_CONFLICT
                - OPERATION_APPROVAL_REQUIRED
                - OPERATION_NOT_APPROVED
                - IDEMPOTENCY_CONFLICT
                - VALIDATION_FAILED
                - RATE_LIMITED
                - INTERNAL_ERROR
                - UPSTREAM_TIMEOUT
            message:
              type: string
            request_id:
              type: string
            resource_id:
              type: string
            resource_status:
              type: string
            retryable:
              type: boolean
            details:
              type: object
            correlation_id:
              type: string
    AssetType:
      type: string
      enum: [knowledge, skill, memory]
    AssetStatus:
      type: string
      enum: [CREATING, PROCESSING, READY, DEGRADED, FAILED, DELETED]
    JobStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELLED, LOST]
    OperationStatus:
      type: string
      enum: [PENDING, APPROVAL_REQUIRED, APPROVED, RUNNING, SUCCEEDED, FAILED, CANCELLED]
    Asset:
      type: object
      properties:
        asset_id:
          type: string
        tenant_id:
          type: string
        type:
          $ref: '#/components/schemas/AssetType'
        name:
          type: string
        status:
          $ref: '#/components/schemas/AssetStatus'
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [type, name, source]
      properties:
        type:
          $ref: '#/components/schemas/AssetType'
        name:
          type: string
        source:
          type: object
        metadata:
          type: object
...
    KnowledgeBase:
      type: object
      properties:
        name:
          type: string
        asset_id:
          type: string
        status:
          $ref: '#/components/schemas/AssetStatus'
        concept_count:
          type: integer
...
    IngestKnowledgeRequest:
      type: object
      required: [name, content, source_type]
      properties:
        name:
          type: string
        content:
          type: string
        source_type:
          type: string
        parser:
          type: string
...
    Skill:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, REVOKED]
        manifest:
          type: object
        created_at:
          type: string
          format: date-time
    RegisterSkillRequest:
      type: object
      required: [manifest, code_package]
      properties:
        manifest:
          type: object
        code_package:
          type: string
          format: binary
...
    Memory:
      type: object
      properties:
        memory_id:
          type: string
        content:
          type: string
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
    AddMemoryRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            type: object
        metadata:
          type: object
...
    JobRun:
      type: object
      properties:
        job_id:
          type: string
        tenant_id:
          type: string
        skill_ref:
          type: string
        status:
          $ref: '#/components/schemas/JobStatus'
        params:
          type: object
        created_at:
          type: string
          format: date-time
    SubmitJobRequest:
      type: object
      required: [skill_ref, inputs]
      properties:
        skill_ref:
          type: string
        inputs:
          type: object
        params:
          type: object
        model_profile:
          type: string
        resource_overrides:
          type: object
...
    Operation:
      type: object
      properties:
        operation_id:
          type: string
        type:
          type: string
        target_resource:
          type: string
        initiator_id:
          type: string
        status:
          $ref: '#/components/schemas/OperationStatus'
        risk_level:
          type: string
          enum: [LOW, MEDIUM, HIGH]
        created_at:
          type: string
          format: date-time
    CreateOperationRequest:
      type: object
      required: [type, target, reason, risk_level]
      properties:
        type:
          type: string
        target:
          type: string
        reason:
          type: string
        risk_level:
          type: string
          enum: [LOW, MEDIUM, HIGH]
    AsyncOperationResponse:
      type: object
      properties:
        operation_id:
          type: string
        resource_id:
          type: string
        status:
          type: string
        message:
          type: string
    ModelDefinition:
      type: object
      properties:
        model_id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [llm, embedding, asr]
    CreateModelRequest:
      type: object
      required: [name, kind]
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [llm, embedding, asr]
        config:
          type: object
...
    SecretRef:
      type: object
      properties:
        ref:
          type: string
        scope:
          type: string
        name:
          type: string
    SecretMetadata:
      type: object
      properties:
        scope:
          type: string
        name:
          type: string
        version:
          type: integer
        created_at:
          type: string
          format: date-time
    CreateSecretRequest:
      type: object
      required: [scope, name, value]
      properties:
        scope:
          type: string
        name:
          type: string
        value:
          type: string
    AuditEvent:
      type: object
      properties:
        event_id:
          type: string
        event_type:
          type: string
        principal:
          type: string
        resource:
          type: string
        action:
          type: string
        result:
          type: string
        details:
          type: object
        created_at:
          type: string
          format: date-time
...
    PageResponseAsset:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Asset'
    PageResponseKnowledgeBase:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/KnowledgeBase'
    PageResponseSkill:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Skill'
    PageResponseMemory:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Memory'
    PageResponseJobRun:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/JobRun'
    PageResponseOperation:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Operation'
...
    PageResponseAuditEvent:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AuditEvent'
    PageResponse:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        has_next:
          type: boolean